state the size bound where a value is narrowed
Two conversions carry a bound the scanner cannot follow: a vector length
bounded by what its offset table can address, and a spec value bounded by a
helper. Each names the destination type's limit at the conversion, which is
what the bound already meant. The size a value reports for itself carries no
such comparison: it is bounded to the SSZ size limit where it enters the size
domain, and that limit is the platform int wherever the int is narrower, so a
guard there would be a branch no input can take.

+++ reflection/unmarshal.go
@@ -748,7 +748,7 @@ func (ctx *ReflectionCtx) unmarshalDynamicVector(targetType *ssztypes.TypeDescri
 	// The offset table holds four bytes per element and its size is computed
 	// in int below, so the length is bounded by what that product can hold.
 	dynVecLen := targetType.Len
-	if dynVecLen > math.MaxInt/4 {
+	if dynVecLen > int64(math.MaxInt)/4 {
 		return sszutils.ErrPlatformOverflowFn("dynamic vector length", targetType.Len)
 	}
